AI/requirements.txt
# NextFit AI API
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
modal

# CatVTON dependencies (compatible with torch 2.1.2, no torch.xpu)
torch==2.1.2
torchvision==0.16.2
diffusers==0.25.1
transformers==4.37.2
accelerate==0.27.2
xformers==0.0.23
matplotlib==3.8.4
numpy==1.26.4
opencv-python==4.10.0.84
Pillow==10.3.0
PyYAML==6.0.1
scipy==1.13.1
setuptools>=51.0.0
scikit-image==0.24.0
tqdm==4.66.4
fvcore==0.1.5.post20221221
cloudpickle==3.0.0
omegaconf==2.3.0
pycocotools==2.0.8
av==12.3.0
peft>=0.17.0
huggingface_hub>=0.34.0,<2.0

# CatVTON / detectron2 / DensePose
iopath>=0.1.7,<0.1.10
yacs>=0.1.8
termcolor>=1.1
tabulate
Ninja==1.11.1.1
